Consume real statistics from inference server

+++ lib/marketplace/model/marketplace_list_data.dart
@@ -17,7 +17,7 @@ class MarketplaceListData {
   double averageRuntime;
   double numberOfSuccessfullRequests;
   int reviews;
-  int pricePerRun;
+  double pricePerRun;
 
   static List<MarketplaceListData> marketplaceList = <MarketplaceListData>[
     MarketplaceListData(
